+++ 1.21.11/Common/src/main/java/fuzs/vehicleupgrade/VehicleUpgrade.java
@@ -18,7 +18,9 @@ import fuzs.vehicleupgrade.handler.*;
 import fuzs.vehicleupgrade.init.ModRegistry;
 import fuzs.vehicleupgrade.network.client.ServerboundOpenEquipmentInventoryMessage;
 import net.minecraft.resources.Identifier;
+import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -58,6 +60,22 @@ public class VehicleUpgrade implements ModConstructor {
         EntityRidingEvents.STOP.register(SprintingMountHandler::onStopRiding);
         PlayerInteractEvents.USE_ENTITY.register(VehicleUpgradeHandler::onUseEntity);
         EntityRidingEvents.START.register(VehicleUpgradeHandler::onStartRiding);
+    }
+
+    @Override
+    public void onCommonSetup() {
+        if (CONFIG.get(CommonConfig.class).speedMobEffectsGrantFlyingSpeed) {
+            MobEffects.SPEED.value()
+                    .addAttributeModifier(Attributes.FLYING_SPEED,
+                            Identifier.withDefaultNamespace("effect.flying_speed"),
+                            0.2F,
+                            AttributeModifier.Operation.ADD_MULTIPLIED_TOTAL);
+            MobEffects.SLOWNESS.value()
+                    .addAttributeModifier(Attributes.FLYING_SPEED,
+                            Identifier.withDefaultNamespace("effect.flying_speed"),
+                            -0.15F,
+                            AttributeModifier.Operation.ADD_MULTIPLIED_TOTAL);
+        }
     }
 
     @Override