+++ 1.21.11/Common/src/main/java/fuzs/vehicleupgrade/config/CommonConfig.java
@@ -4,15 +4,16 @@ import fuzs.puzzleslib.api.config.v3.Config;
 import fuzs.puzzleslib.api.config.v3.ConfigCore;
 
 public class CommonConfig implements ConfigCore {
-    @Config(
-            description = "Prevent a slower block breaking speed caused by the player not standing on the ground while riding any vehicle.",
-            gameRestart = true
-    )
+    @Config(description = "Prevent a slower block breaking speed caused by the player not standing on the ground while riding any vehicle.",
+            gameRestart = true)
     public boolean removePassengerMiningSpeedMalus = true;
-    @Config(
-            category = ServerConfig.HORSE_CATEGORY,
+    @Config(category = ServerConfig.HORSE_CATEGORY,
             description = "Set the step height for all horses above one block to be able to run up from path blocks, farmland, etc.",
-            gameRestart = true
-    )
+            gameRestart = true)
     public boolean increaseHorseStepHeight = true;
+    @Config(description = {
+            "Makes speed and slowness mob effects provide flying speed in addition to movement speed.",
+            "This allows mobs such as the happy ghast and bees to benefit from these effects."
+    }, gameRestart = true)
+    public boolean speedMobEffectsGrantFlyingSpeed = true;
 }