+++ 1.21.11/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.11.1-1.21.11] - 2025-12-24
+
+### Added
+
+- Speed and slowness mob effects now provide flying speed in addition to movement speed
+- This allows mobs such as the happy ghast and bees to benefit from these effects
+
 ## [v21.11.0-1.21.11] - 2025-12-19
 
 ### Changed